requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
colorama==0.4.6